+++ CHANGELOG.md
@@ -184,3 +184,4 @@
 * Rename `Group`'s method from `inverse` to `invert`
 * Remove `not` method of `Boolean`, (use `invert` instead)
 * Remove `Injective` and `Has` type families
+* Change superclasseses for `Liftable` and `Lowerable` classes to provide a law